atualizar readme e colocar alguns comments
