+++ Atividades-Avaliativas/Entrega-N1/mergesort/README.md
@@ -1,0 +1,35 @@
+Claro, vou explicar o código linha por linha:
+
+1. `#include <stdio.h>`: Isso inclui a biblioteca de entrada/saída padrão em C, que é necessária para funções como `printf` e `scanf`.
+
+2. `#include <stdlib.h>`: Isso inclui a biblioteca padrão de C, que é necessária para funções como `malloc` e `rand`.
+
+3. `#include <time.h>`: Isso inclui a biblioteca de tempo em C, que é usada para medir o tempo de execução do algoritmo.
+
+4. `#define TAMANHO 10000`: Isso define uma constante chamada `TAMANHO` com o valor 10000. Essa constante será usada para determinar o tamanho do array a ser ordenado.
+
+5. `void geraNumero(int *vet, int op)`: Isso é um protótipo de função. Ele declara uma função chamada `geraNumero` que recebe um ponteiro para um array de inteiros (`int *vet`) e um inteiro `op` como argumentos. Essa função será usada para gerar diferentes tipos de arranjos (ordenados, invertidos ou aleatórios).
+
+6. `void merge(int arr[], int left, int middle, int right, int *qtd_trocas, int *qtd_comparacoes)`: Este é o protótipo de outra função chamada `merge`, que será usada para mesclar duas partes de um array durante o algoritmo de ordenação Merge Sort. Ela recebe como argumentos o array a ser ordenado (`arr`), os índices de início (`left`), meio (`middle`) e fim (`right`) das partes a serem mescladas, e dois ponteiros para inteiros que rastreiam a quantidade de trocas e comparações.
+
+7. `void mergeSort(int arr[], int left, int right, int *qtd_trocas, int *qtd_comparacoes)`: Este é o protótipo da função `mergeSort`, que é a implementação principal do algoritmo de ordenação Merge Sort. Recebe o array a ser ordenado, bem como os índices de início e fim da parte do array a ser ordenada, além dos ponteiros para rastrear as trocas e comparações.
+
+As linhas de código após os protótipos de função são as implementações dessas funções. Vou explicar o código principal no `main`:
+
+8. `int arr[TAMANHO]`: Isso declara um array de inteiros chamado `arr` com tamanho definido pela constante `TAMANHO`. Este é o array que será ordenado.
+
+9. `int arr_size = sizeof(arr) / sizeof(arr[0])`: Isso calcula o tamanho do array `arr` dividindo o tamanho total do array pelo tamanho de um elemento individual no array. Isso é usado para determinar o número de elementos no array.
+
+10. `int qtd_trocas = 0;` e `int qtd_comparacoes = 0;`: Essas linhas declaram duas variáveis inteiras que serão usadas para rastrear o número de trocas e comparações realizadas durante o processo de ordenação.
+
+11. `geraNumero(arr, 3);`: Esta linha chama a função `geraNumero` para preencher o array `arr` com números aleatórios.
+
+12. `float tempo_inicial = clock();`: Isso captura o tempo de início da execução do algoritmo, usando a função `clock()` da biblioteca `time.h`.
+
+13. `mergeSort(arr, 0, arr_size - 1, &qtd_trocas, &qtd_comparacoes);`: Esta linha chama a função `mergeSort` para ordenar o array `arr`. Os argumentos são o array, o índice de início (0), o índice final (`arr_size - 1`), bem como os ponteiros para rastrear as trocas e comparações.
+
+14. `float tempo_final = clock() - tempo_inicial;`: Isso calcula o tempo total de execução subtraindo o tempo de início do tempo atual.
+
+15. As linhas finais do código imprimem o número de comparações, trocas e o tempo de execução do algoritmo Merge Sort.
+
+Esse código em C implementa o algoritmo de ordenação Merge Sort e o utiliza para ordenar um array de inteiros gerado aleatoriamente. Ele também rastreia a quantidade de comparações, trocas e o tempo de execução do algoritmo.
