+++ Atividades-Avaliativas/Entrega-N1/quicksort/README.md
@@ -1,0 +1,81 @@
+Este código é um programa em C que implementa o algoritmo de ordenação Quick Sort e mede o número de comparações, trocas e o tempo de execução para ordenar um vetor de inteiros. Vou explicar o código linha por linha:
+
+```c
+#include <stdio.h>
+#include <stdlib.h>
+#include <time.h>
+```
+- As primeiras três linhas são diretivas de pré-processador que incluem as bibliotecas necessárias para o programa. `stdio.h` para entrada e saída padrão, `stdlib.h` para funções relacionadas à alocação de memória, e `time.h` para medir o tempo de execução.
+
+```c
+#define TAMANHO 10000
+```
+- Esta linha define uma constante chamada `TAMANHO` com o valor 10000, que representa o tamanho do vetor a ser ordenado.
+
+```c
+void geraNumero(int *vet, int op) {
+    // ...
+}
+```
+- Aqui começa a definição da função `geraNumero`, que é responsável por preencher o vetor com números de acordo com a opção escolhida. Ela recebe um ponteiro para o vetor e um inteiro que representa a opção (1 para números ordenados, 2 para números inversamente ordenados, 3 para números aleatórios).
+
+```c
+void troca(int arr[], int i, int j, int *qtd_trocas) {
+    // ...
+}
+```
+- Esta função `troca` é responsável por trocar dois elementos de um vetor. Ela recebe o vetor, dois índices `i` e `j` e um ponteiro para um contador de trocas `qtd_trocas`.
+
+```c
+int particao(int x[], int p, int r, int *qtd_comparacoes, int *qtd_trocas) {
+    // ...
+}
+```
+- A função `particao` é uma parte fundamental do algoritmo Quick Sort. Ela recebe o vetor, dois índices `p` e `r`, e dois ponteiros para contadores de comparações e trocas. Esta função encontra um pivô e rearranja o vetor de forma que elementos menores que o pivô estejam à esquerda e elementos maiores estejam à direita.
+
+```c
+void quicksort(int arr[], int p, int r, int *qtd_comparacoes, int *qtd_trocas) {
+    // ...
+}
+```
+- A função `quicksort` é a função principal que implementa o algoritmo Quick Sort. Ela recebe o vetor, os índices de início `p` e fim `r` do subvetor a ser ordenado, e os ponteiros para contadores de comparações e trocas. Ela chama a função `particao` para dividir o vetor e recursivamente ordena as duas metades.
+
+```c
+int main() {
+    int arr[TAMANHO];
+    int arr_size = sizeof(arr) / sizeof(arr[0]);
+    int qtd_trocas = 0;
+    int qtd_comparacoes = 0;
+```
+- A função `main` é a função principal do programa. Ela declara um vetor `arr` de tamanho `TAMANHO`, calcula o tamanho do vetor em `arr_size`, e inicializa contadores de trocas e comparações.
+
+```c
+    geraNumero(arr, 1);
+```
+- Aqui, a função `geraNumero` é chamada para preencher o vetor `arr` com números ordenados.
+
+```c
+    float tempo_inicial = clock();
+```
+- Esta linha registra o tempo de início da execução do algoritmo usando a função `clock()`.
+
+```c
+    quicksort(arr, 0, arr_size - 1, &qtd_comparacoes, &qtd_trocas);
+```
+- Chama a função `quicksort` para ordenar o vetor `arr`.
+
+```c
+    float tempo_final = clock() - tempo_inicial;
+```
+- Calcula o tempo de execução subtraindo o tempo de início do tempo atual.
+
+```c
+    printf("Algoritmo de ordenacao Quick Sort: \n");
+    printf("\nQuantidade de comparacoes: %d\n", qtd_comparacoes);
+    printf("Quantidade de trocas: %d\n", qtd_trocas);
+    printf("Tempo de execucao do algoritmo: %.3f\n", tempo_final / 1000);
+
+    return 0;
+}
+```
+- Finalmente, o programa exibe informações sobre o algoritmo de ordenação, o número de comparações, o número de trocas e o tempo de execução em segundos.
